+++ projects/addon-doc/src/tokens/example-content-processor.ts
@@ -1,0 +1,8 @@
+import {InjectionToken} from '@angular/core';
+import {identity, TuiHandler} from '@taiga-ui/cdk';
+
+export const TUI_DOC_EXAMPLE_CONTENT_PROCESSOR = new InjectionToken<
+    TuiHandler<Record<string, string>, Record<string, string>>
+>('Processes content in example', {
+    factory: () => identity,
+});