+++ projects/core/tokens/index.ts
@@ -1,4 +1,3 @@
-export * from './checkbox-appearance';
 export * from './data-list-accessor';
 export * from './data-list-host';
 export * from './document-or-shadow-root';
@@ -9,6 +8,7 @@ export * from './icons';
 export * from './icons-path';
 export * from './mode';
 export * from './option-content';
+export * from './sanitizer';
 export * from './scroll-ref';
 export * from './textfield-appearance';
 export * from './value-accessor';