+++ projects/core/types/dialog-size.ts
@@ -1,3 +1,3 @@
 import {TuiSizeL, TuiSizeS} from './size';
 
-export type TuiDialogSize = TuiSizeS | TuiSizeL | 'fullscreen';
+export type TuiDialogSize = TuiSizeS | TuiSizeL | 'fullscreen' | 'page';