+++ projects/demo/src/modules/app/app.component.ts
@@ -1,4 +1,4 @@
-import {ChangeDetectorRef, Component, HostBinding, Inject} from '@angular/core';
+import {ChangeDetectorRef, Component, Inject} from '@angular/core';
 import {Router} from '@angular/router';
 import {TUI_IS_ANDROID, TUI_IS_IOS, tuiPure, watch} from '@taiga-ui/cdk';
 import {VERSION} from '@taiga-ui/core';
@@ -13,7 +13,6 @@ import {changeDetection} from '../../change-detection-strategy';
     changeDetection,
 })
 export class AppComponent {
-    @HostBinding('class._landing')
     landing = false;
 
     readonly version = VERSION;