+++ projects/demo/src/modules/app/app.providers.ts
@@ -1,6 +1,7 @@
 import {LocationStrategy, PathLocationStrategy} from '@angular/common';
 import {Title} from '@angular/platform-browser';
 import {
+    TUI_DOC_CODE_EDITOR,
     TUI_DOC_DEFAULT_TABS,
     TUI_DOC_LOGO,
     TUI_DOC_PAGES,
@@ -11,10 +12,13 @@ import {TUI_SANITIZER} from '@taiga-ui/cdk';
 import {iconsPathFactory, TUI_ICONS_PATH} from '@taiga-ui/core';
 import {NgDompurifySanitizer} from '@tinkoff/ng-dompurify';
 import {HIGHLIGHT_OPTIONS} from 'ngx-highlightjs';
+import {TUI_DOC_EXAMPLE_CONTENT_PROCESSOR} from '../../../../addon-doc/src/tokens/example-content-processor';
 import {PROMPT_PROVIDER} from '../customization/dialogs/examples/prompt/prompt.component';
+import {FrontEndExample} from '../interfaces/front-end-example';
 import {SEE_ALSO_GROUPS} from './app.const';
 import {LOGO_CONTENT} from './logo/logo.component';
 import {pages} from './pages';
+import {StackblitzService} from './stackblitz/stackblitz.service';
 
 export const DEFAULT_TABS = [
     $localize`Description and examples`,
@@ -33,6 +37,45 @@ export const HIGHLIGHT_OPTIONS_VALUE = {
         xml: () => import('highlight.js/lib/languages/xml'),
     },
 };
+
+export function exampleContentProcessor(content: FrontEndExample): FrontEndExample {
+    if (!content.TypeScript) {
+        return content;
+    }
+
+    const withChangeDetectionImport = addIntoExistingImport(
+        content.TypeScript,
+        'ChangeDetectionStrategy',
+        '@angular/core',
+    );
+
+    return {
+        ...content,
+        TypeScript: withChangeDetectionImport
+            .replace(/import {encapsulation} from '..\/.*';\n/gm, '')
+            .replace(/import {changeDetection} from '..\/.*';\n/gm, '')
+            .replace(/\n +encapsulation,/gm, '')
+            .replace(
+                /changeDetection,/gm,
+                'changeDetection: ChangeDetectionStrategy.OnPush,',
+            ),
+    };
+}
+
+export function addIntoExistingImport(
+    data: string = '',
+    entity: string = '',
+    packageName: string = '',
+): string {
+    const packageImportsRegex = new RegExp(
+        `(?:import\\s?\\{\\r?\\n?)(?:(?:.*),\\r?\\n?)*?(?:.*?)\\r?\\n?} from (?:'|")${packageName}(?:'|");`,
+        'gm',
+    );
+
+    return data.replace(packageImportsRegex, parsed => {
+        return parsed.replace('{', `{${entity}, `);
+    });
+}
 
 export const ICONS_PATH = iconsPathFactory('assets/taiga-ui/icons/');
 
@@ -74,5 +117,13 @@ export const APP_PROVIDERS = [
     {
         provide: TUI_DOC_LOGO,
         useValue: LOGO_CONTENT,
+    },
+    {
+        provide: TUI_DOC_CODE_EDITOR,
+        useClass: StackblitzService,
+    },
+    {
+        provide: TUI_DOC_EXAMPLE_CONTENT_PROCESSOR,
+        useValue: exampleContentProcessor,
     },
 ];