+++ projects/demo/src/modules/components/button/button.component.ts
@@ -1,6 +1,7 @@
 import {default as example1Html} from '!!raw-loader!./examples/1/index.html';
 import {default as example1Ts} from '!!raw-loader!./examples/1/index.ts';
 import {default as example2Html} from '!!raw-loader!./examples/2/index.html';
+import {default as example2Ts} from '!!raw-loader!./examples/2/index.ts';
 import {default as example3Html} from '!!raw-loader!./examples/3/index.html';
 import {default as example3Less} from '!!raw-loader!./examples/3/index.less';
 import {default as example3Ts} from '!!raw-loader!./examples/3/index.ts';
@@ -40,6 +41,7 @@ export class ExampleTuiButtonComponent extends AbstractExampleTuiField {
 
     readonly example2: FrontEndExample = {
         HTML: example2Html,
+        TypeScript: example2Ts,
     };
 
     readonly example3: FrontEndExample = {