+++ projects/demo/src/modules/customization/i18n/i18n.component.ts
@@ -1,3 +1,4 @@
+import {default as readme} from '!!raw-loader!../../../../../i18n/README.md';
 import {Component, ViewEncapsulation} from '@angular/core';
 import {changeDetection} from '../../../change-detection-strategy';
 
@@ -8,4 +9,8 @@ import {changeDetection} from '../../../change-detection-strategy';
     encapsulation: ViewEncapsulation.None,
     changeDetection,
 })
-export class I18nComponent {}
+export class I18nComponent {
+    readonly supported = readme
+        .split('> A package with tools for Taiga UI library i18n')[1]
+        .split(`It's a part of [**Taiga UI**]`)[0];
+}