+++ projects/demo/src/modules/customization/icon-set/icon-set.component.ts
@@ -1,5 +1,4 @@
 import {default as example1Html} from '!!raw-loader!./examples/1/index.html';
-import {default as example1Less} from '!!raw-loader!./examples/1/index.less';
 import {default as example1Ts} from '!!raw-loader!./examples/1/index.ts';
 
 import {Component} from '@angular/core';
@@ -17,7 +16,6 @@ export class IconSetComponent {
     readonly example1: FrontEndExample = {
         TypeScript: example1Ts,
         HTML: example1Html,
-        LESS: example1Less,
     };
 
     readonly names = [...Object.keys(tuiCoreIcons), ...Object.keys(tuiKitIcons)];