+++ projects/demo/src/modules/services/dialogs/dialogs.component.ts
@@ -30,7 +30,7 @@ import {FrontEndExample} from '../../interfaces/front-end-example';
 const TOKEN = `{
     provide: TUI_DIALOGS_CLOSE,
     deps: [Router],
-    useFactory: ({events}) => events.pipe(filter(event => event instanceof NavigationStart)),
+    useFactory: ({events}) => events,
 }`;
 
 @Component({
@@ -84,7 +84,13 @@ export class ExampleTuiDialogsComponent {
 
     required = false;
 
-    readonly sizeVariants: ReadonlyArray<TuiDialogSize> = ['s', 'm', 'l', 'fullscreen'];
+    readonly sizeVariants: ReadonlyArray<TuiDialogSize> = [
+        's',
+        'm',
+        'l',
+        'fullscreen',
+        'page',
+    ];
 
     size: TuiDialogSize = this.sizeVariants[1];