+++ projects/i18n/languages/turkish/core.ts
@@ -1,0 +1,23 @@
+import {LanguageCore} from '@taiga-ui/i18n/interfaces';
+
+export const TUI_TURKISH_LANGUAGE_CORE: LanguageCore = {
+    months: [
+        'Ocak',
+        'Şubat',
+        'Mart',
+        'Nisan',
+        'Mayıs',
+        'Haziran',
+        'Temmuz',
+        'Ağustos',
+        'Eylül',
+        'Ekim',
+        'Kasım',
+        'Aralık',
+    ],
+    close: 'Kapat',
+    nothingFoundMessage: 'Kayıt Bulunamadı',
+    defaultErrorMessage: 'Değer hatalı',
+    spinTexts: ['Geri', 'İleri'],
+    shortWeekDays: ['Ptsi', 'Salı', 'Çrş', 'Prş', 'Cma', 'Ctsi', 'Pzr'],
+};