+++ projects/kit/components/avatar/avatar.module.ts
@@ -1,9 +1,10 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
+import {TuiLazyLoadingModule} from '@taiga-ui/kit/directives';
 import {TuiAvatarComponent} from './avatar.component';
 
 @NgModule({
-    imports: [CommonModule],
+    imports: [CommonModule, TuiLazyLoadingModule],
     declarations: [TuiAvatarComponent],
     exports: [TuiAvatarComponent],
 })